LayoutTests/platform/qt/fast/dom/Window/window-resize-contents-expected.txt

CONSOLE MESSAGE: line 19: Initial reference node dimensions 800 x 600
CONSOLE MESSAGE: line 20: Increasing window size by 10 x 10
CONSOLE MESSAGE: line 23: Post-resize reference node dimensions 800 x 600
This test checks that the yellow reference DOM node (which should be as big as the window) gets resized when the window is resized.

To avoid relayouts and repaints caused by DOM-based logging, it doesn't output anything. Please check the console for confirmation that the node dimensions increase by 10x10.
